Andy's commit 7 1. Renamed the package to csc207project_code, so we can use it inside. 2. Moved Main.java to csc207project_code package, so it is easier for user to find it. 3. Added a package called csc207project_test, so we can have a test file for each class separately. 4. Moved test in GameTest.java to the csc207project_test, in TestInitializeBoard.java and TestChess.java.
