+++ csc207project_code/Chesses/Elephant.java
