+++ csc207project_code/Chesses/Soldier.java
