+++ csc207project_code/Model/getNextPosition.java
