+++ csc207project_test/Model/TestChess.java
@@ -1,22 +1,16 @@
-import Chesses.*;
-import Control.*;
-import Model.*;
+package Model;
+
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
 
+import Chesses.Advisor;
 
-public class GameTest {
+
+public class TestChess {
     @Test
     void testGetLocation() {
         Location expectedLocation = new Location(9, 3);
         Advisor advisor1 = new Advisor(expectedLocation, "Red", "Advisor");
         assertEquals(expectedLocation, advisor1.getLocation());
     }
-    @Test
-    void testInitializeBoard(){
-        InitializeBoard initializeBoard = new InitializeBoard();
-        System.out.println(initializeBoard.getBoard());
-    }
-
-
 }
