+++ csc207project_test/TestInitializeBoard.java
@@ -1,0 +1,16 @@
+import org.junit.jupiter.api.Test;
+import static org.junit.jupiter.api.Assertions.*;
+
+
+public class TestInitializeBoard {
+    @Test
+    void testInitializeBoard(){
+        InitializeBoard initializeBoard = new InitializeBoard();
+        System.out.println(initializeBoard.getBoard());
+    }
+    @Test
+    void testGetBoard(){
+        InitializeBoard initializeBoard = new InitializeBoard();
+        System.out.println(initializeBoard.getBoard());
+    }
+}
